move ForeachFuncInfo to its own file
